Завершено написание криптографии для BAC Созданы класс хэширования SHA-1 и класс шифрования DES3 Функции криптографии убраны в пространство имён Crypto В этом пространстве имён добавлена функция для генерации байтов с младшим битом чётности

+++ BAC_reader/CardReader.hpp
@@ -50,7 +50,7 @@ public:
 			return {};
 		}
 		while (*pReader != '\0') {
-			int readerNameLen = wcslen(pReader);
+			size_t readerNameLen = wcslen(pReader);
 
 			/* Производим копирование названия во временную строку */
 			wstring tempString = L"";
